투명 적용
투명 적용

+++ DX3D_2412/Framework/Math/GameMath.cpp
@@ -3,7 +3,7 @@
 
 int GameMath::Random(const int& min, const int& max)
 {
-    return rand() % (max - min) + min;
+    return (int)rand() % (max - min) + min;
 }
 
 float GameMath::Random(const float& min, const float& max)
